requirements.txt
pygame==2.6.1
pillow==11.1.0
requests==2.28.2